requirements.txt
streamlit
requests
pandas
AgGrid
GridOptionsBuilder